Set PREFIX if it's blank
If INTBASIC.SYSTEM is invoked with an empty path, most ProDOS calls
(e.g. CAT) are sad. A repro is to launch INTBASIC.SYSTEM from
BASIC.SYSTEM.

Resolve this by testing if a PREFIX is set. If not, use DEVNUM and
ON_LINE to set the PREFIX to something likely.