Add support for Programmer's Aid Music Subroutine
A copy of the subroutine is include - this consumes a bit more memory
but HIMEM is still higher than $9600 so that's fine. Then CALL is
intercepted and a specific call to $D717 is rerouted to our copy of
the subroutine.

(In the future, the routine could be placed in LCBank2 $D400-$DFFF
space which is mostly unused by ProDOS.)

Fixes #12